Refactor structs into submodules.
